address coderabbit and copilot suggestions

+++ lib/GfxRenderer/GfxRenderer.h
@@ -195,8 +195,8 @@ class GfxRenderer {
   // same rectangle to writeFramebufferRegion to restore the saved pixels.
   // Enables partial-repaint patterns (e.g. moving a selection highlight)
   // without re-rendering the whole page.
-  size_t readFramebufferRegion(uint16_t x, uint16_t y, uint16_t w, uint16_t h, uint8_t* dst, size_t dstCapacity) const;
-  void writeFramebufferRegion(uint16_t x, uint16_t y, uint16_t w, uint16_t h, const uint8_t* src);
+  size_t readFramebufferRegion(int x, int y, int w, int h, uint8_t* dst, size_t dstCapacity) const;
+  void writeFramebufferRegion(int x, int y, int w, int h, const uint8_t* src);
 
   // Text
   int getTextWidth(int fontId, const char* text, EpdFontFamily::Style style = EpdFontFamily::REGULAR,
